fix(ci): CHARMHUB_TOKEN -> CHARMCRAFT_AUTH
Incorrect secret name `CHARMHUB_TOKEN` is being used within the CI definition. This incorrect name causes the CI to fail when checking the charm library versions because GitHub cannot not find the secret resource. Setting the name to `CHARMCRAFT_AUTH` will fix the issue since the secret name is configured to be `CHARMCRAFT_AUTH`.

Signed-off-by: Jason C. Nucciarone <nuccitheboss@ubuntu.com>